# Optimize performance when comparing big enum-case unions