This change prevents the previous recursion which made comparisons against object types slow.

call graph before this PR:

https://blackfire.io/profiles/ad03e612-d5c9-410b-9bc4-763dfaa3d66b/graph

<img width="962" alt="grafik" src="https://github.com/phpstan/phpstan-src/assets/120441/1ec80acb-2b62-457c-8b7d-d02665e70a11">

closes https://github.com/phpstan/phpstan/issues/10979

Before this PR
```
php bin/phpstan analyze --debug LibraryMasterEnum.php
  23.67s user 0.14s system 99% cpu 23.858 total
```

After this PR
```
php bin/phpstan analyze --debug LibraryMasterEnum.php
  4.11s user 0.15s system 99% cpu 4.291 total
```